Move user agent service files to their own directory, as they're only ever run together in their own process. r=mossop
